requirements.txt
boto3==1.23.2
python-dotenv==0.20.0
lxml
mechanize
requests
